Makes clik a package instead of a single module.
I ran into an odd bug with find_packages(). (Read: I probably
misunderstand what it's doing.) I tried using `exclude=['test']`
but the test code was still ending up in the built distribution.

A prune in the MANIFEST.in file does what I want (namely, leaving
test code out of the built package). It doesn't feel "right" but
it works and it's not actually all that ugly so meh, just gonna
take that as a win and move on.