Lots of tweaks based on code review, and one bugfix.
Bugfix: I had always intended that run_children() returned the
child exit codes. The test was extended to demonstrate that fact,
and it failed because it wasn't implemented! Adding a `return`
fixed this.

The rest of the changes are mostly sweeping up stuff that isn't
used and tweaking some wording for clarity.