Enables testing in all supported Python versions.
A small difference in error messages aside, the core concept
seems to work pretty well as-is across all supported versions.
(Haven't tested pypy3 yet since I don't have it on this machine.)

Of course, this is still missing aliases and command handling,
which are the two main differences between 2 and 3. So, while
it's a nice step forward, there's still a long way to go.