Distills functional tests into their smallest form.
I intend to point to this in the documentation as a place to look
if you already know Python well and just want to see what the
thing does. To that end, I thought it'd be lower overhead for the
reader to have absolutely no extraneous stuff in those files.