Adds alias functional test, re-enables aliases.
Things are still broken on Python 2. The current plan is to get
all the things working in the "reference" Python 3 implementation,
then figure out how to support Python 2 without hurting the
cleanliness of the 3 implementation. (Then try to make the code
and internal API nice, then work on testing the crap out of it.
Then documentation. And I'm sure a lot more stuff along the way
that I'm not thinking about right now.)