Cleans up terminology and notation.
"Command" has been renamed to "bare" to be less confusing.

"rv" and "rvs" (return value) have been renamed to "ec" and
"ecs" (exit code) to be consistent with shell terminology and
other clik documentation.

Two changes to args._clik_cmd: (1) it is now called the
"command stack" or the "stack" to be less confusing and (2) the
name of the dest is now specified in the STACK module-level
variable (and is currently set to '_clik_stack').

A known-broken/incomplete but maybe closer implementation for
bare destination argument handling is also in here.