Disables mutex groups and posargs for bare commands.
These never worked in the first place and were never intended
to be supported. This commit makes that official.

I don't know whether it's possible to support these, but I'm so
burnt out on thinking about how to bend argparse to do all the
things I want that I've been avoiding this project.

So we'll say: maybe someday. But for now we'll move forward.
And it's always easier to add features than remove ones (even
if they're broken) right?