Adds a test for invalid subcommand name.
Note on the tox.ini change: Certain PYTHONHASHSEED values were
causing the error message output to be in a different order.
Because I don't actually care about that I just disabled the
hash randomization. (I realize this could mask other Python 2.6
hash seed related bugs later on, but if that happens, it happens.)