Major reordering of functional test cases.
Eventually I want to point to the functional tests as a way
for Python/argparse experts to quickly see what clik has to
offer. (You know, when there's documentation.)

As for the PYTHONHASHSEED change -- it's ugly. I just kept
trying different values to make the py26 error messages come
out in the "correct" order. If there's another problem of this
kind, marc might have to be modified to do glob/regex matching
on lines. (For this case, it was quicker and easier to find a
friendly hash seed; that is not a solution that's going to keep
working.)