Fixes linter violations in compatibility code.
I see no reason to formally document compatibility code. And even if we did,
I can't imagine trying to generate the documentation for it. Finally, in a
technical sense, it's just a hack to keep us limping along until we get to
the "ideal, final" stage of the project where everyone is on 3.6 and we can
kill all the compat stuff.

Or I might be lazy.

Or both.