Updates dependencies to their latest versions.
* coverage (4.4.1 -> 4.4.2)
* flake8-commas (0.4.3 -> 1.0.0)
* flake8-debugger (1.4.0 -> 3.0.0)
* flake8-docstrings (1.1.0 -> 1.3.0)
* pytest (3.2.3 -> 3.2.5)
* sphinx (1.6.5 -> 1.6.6)

Note that flake8-import-order has been held back since it is
causing linting errors that will need to be fixed.

Also note that pytest has a 3.3 release available. 3.3 drops
support for Python 2.6, which clik still officially has. So
for now, pytest has been updated to the latest release that
supports 2.6, and we'll figure out later what to do moving
forward.